rename selfChange -> addressChange as that's what it is now
